requirements.txt
pyperclip
keyboard
win10toast
zhipuai
python-dotenv